Break out if_this_then_that_helpers.py from ifttt_gmail.py
Fix issues where examples/if_this_then_that/ifttt_gmail.py code can't find flask_helpers or images/ in examples/
